+++ src/main/java/racingcar/dto/CarResponse.java
@@ -1,4 +1,4 @@
-package racingcar.controller.dto;
+package racingcar.dto;
 
 import racingcar.domain.Car;
