+++ src/main/java/racingcar/dto/PlaysResponse.java
@@ -1,4 +1,4 @@
-package racingcar.controller.dto;
+package racingcar.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
