+++ src/test/java/racingcar/controller/RacingCarControllerTest.java
@@ -17,9 +17,9 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
-import racingcar.controller.dto.PlaysRequest;
-import racingcar.controller.dto.PlaysResponse;
 import racingcar.domain.Car;
+import racingcar.dto.PlaysRequest;
+import racingcar.dto.PlaysResponse;
 import racingcar.service.RacingCarService;
 import racingcar.utils.NumberGenerator;
 import racingcar.utils.RandomNumberGenerator;
